web_ui/src/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0

# TODO: 실제 AI 모델 사용 시 아래 주석을 해제하고 설치하세요
# gfpgan
# basicsr
# facexlib
# realesrgan
# codeformer
# torch
# torchvision